fix: handle lines with Unicode characters
The null-ls diagnostic reports the wrong range for the CSpell error if
the line contains a Unicode character.

As a workaround, we read the misspelled word from the diagnostic's
user_data. And only use the word from the range to trigger a new
diagnostic.

That means that we will still see the wrong range in the diagnostic, but
the code actions will work as expected.

See: https://github.com/jose-elias-alvarez/null-ls.nvim/issues/1630