fix(code actions): remove asterisk from fixes
The list of words returned by the CSpell CLI, when called with the
`--show-suggestions` flag, includes first the fixes and then the
suggestions.

The fixes are marked with an asterisk to differentiate them from the
suggestions.

This commit removes the asterisk and makes the distinction between
fixes and suggestions.

The order was already wrong, but that's been addressed upstream, in the
none-ls.nvim repo.

Resolves #47